MultiSpeaker almost working

+++ sources/AppBundle/CFP/PhotoStorage.php
@@ -39,10 +39,12 @@ class PhotoStorage
         }
 
         // delete all formats
-        foreach (self::FORMAT as $format => $sizes) {
-            $files = glob($this->basePath . '/' . $speaker->getEventId() . '/' . $format . '/' . $speaker->getId() . '.*');
-            foreach ($files as $file) {
-                unlink($file);
+        if ($speaker->getId() === null) {
+            foreach (self::FORMAT as $format => $sizes) {
+                $files = glob($this->basePath . '/' . $speaker->getEventId() . '/' . $format . '/' . $speaker->getId() . '.*');
+                foreach ($files as $file) {
+                    unlink($file);
+                }
             }
         }
